+++ lib/qrcode_reader_web.dart
@@ -1,7 +1,5 @@
 library qrcode_reader_web;
 
-export 'src/qrcode_reader.dart';
 export 'src/exceptions/qrcode_reader_exception.dart';
 export 'src/objects/qrcode_capture.dart';
 export 'src/ui/qrcode_reader_widget.dart';
-export 'src/ui/qrcode_reader_controller.dart';