+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.2
+
+* Fixed an issue when building for platforms other than Web that used the QRCodeWidget.
+* Added conditional import to ensure the build is only for Web; otherwise, it will return a SizedBox.
+
 ## 1.0.1
 
 * Create Flutter Web Plugin.